Store transport keys and certificates in a single shared secret.
This facilitates a move to StatefulSets where the mounted secrets must be the same between
all the Pods in the same StatefulSet
