+++ operators/pkg/controller/elasticsearch/certificates/ca_reconcile.go
@@ -107,7 +107,7 @@ func Reconcile(
 	}
 
 	// reconcile transport certificates
-	result, err := transport.ReconcileTransportCertificateSecrets(
+	result, err := transport.ReconcileTransportCertificatesSecrets(
 		c,
 		scheme,
 		transportCA,
