+++ operators/pkg/controller/elasticsearch/certificates/transport/pod_secret_test.go
@@ -6,121 +6,219 @@ package transport
 
 import (
 	"testing"
+	"time"
 
-	"github.com/elastic/cloud-on-k8s/operators/pkg/apis/elasticsearch/v1alpha1"
-	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/label"
-	"github.com/elastic/cloud-on-k8s/operators/pkg/utils/k8s"
+	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/certificates"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
-func TestEnsureTransportCertificateSecretExists(t *testing.T) {
-	es := v1alpha1.Elasticsearch{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "test-es",
-		},
-	}
-
-	pod := corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "pod",
-		},
-	}
-
-	defaultPodSecret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "pod-certs",
-			Labels: map[string]string{
-				LabelCertificateType:       LabelCertificateTypeTransport,
-				label.PodNameLabelName:     pod.Name,
-				label.ClusterNameLabelName: es.Name,
-			},
-		},
-	}
-
-	defaultPodSecretWith := func(setter func(secret *corev1.Secret)) *corev1.Secret {
-		secret := defaultPodSecret.DeepCopy()
-		setter(secret)
-		return secret
-	}
-
+func Test_shouldIssueNewCertificate(t *testing.T) {
 	type args struct {
-		c      k8s.Client
-		scheme *runtime.Scheme
-		owner  v1alpha1.Elasticsearch
-		pod    corev1.Pod
-		labels map[string]string
+		secret       corev1.Secret
+		pod          *corev1.Pod
+		rotateBefore time.Duration
 	}
 	tests := []struct {
-		name    string
-		args    args
-		want    func(*testing.T, *corev1.Secret)
-		wantErr bool
+		name string
+		args args
+		want bool
 	}{
 		{
-			name: "should create a secret if it does not already exist",
+			name: "missing cert in secret",
 			args: args{
-				c:     k8s.WrapClient(fake.NewFakeClient()),
-				owner: es,
-				pod:   pod,
+				secret:       corev1.Secret{},
+				rotateBefore: certificates.DefaultRotateBefore,
 			},
-			want: func(t *testing.T, secret *corev1.Secret) {
-				// owner references are set upon creation, so ignore for comparison
-				expected := defaultPodSecretWith(func(s *corev1.Secret) {
-					s.OwnerReferences = secret.OwnerReferences
-				})
-				assert.Equal(t, expected, secret)
+			want: true,
+		},
+		{
+			name: "invalid cert data",
+			args: args{
+				secret: corev1.Secret{
+					Data: map[string][]byte{
+						PodCertFileName(testPod.Name): []byte("invalid"),
+					},
+				},
+				rotateBefore: certificates.DefaultRotateBefore,
+			},
+			want: true,
+		},
+		{
+			name: "pod name mismatch",
+			args: args{
+				secret: corev1.Secret{
+					Data: map[string][]byte{
+						PodCertFileName(testPod.Name): pemCert,
+					},
+				},
+				pod:          &corev1.Pod{ObjectMeta: metav1.ObjectMeta{Name: "different"}},
+				rotateBefore: certificates.DefaultRotateBefore,
+			},
+			want: true,
+		},
+		{
+			name: "valid cert",
+			args: args{
+				secret: corev1.Secret{
+					Data: map[string][]byte{
+						PodCertFileName(testPod.Name): pemCert,
+					},
+				},
+				rotateBefore: certificates.DefaultRotateBefore,
+			},
+			want: false,
+		},
+		{
+			name: "should be rotated soon",
+			args: args{
+				secret: corev1.Secret{
+					Data: map[string][]byte{
+						PodCertFileName(testPod.Name): pemCert,
+					},
+				},
+				rotateBefore: certificates.DefaultCertValidity, // rotate before the same duration as total validity
+			},
+			want: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if tt.args.pod == nil {
+				tt.args.pod = &testPod
+			}
+
+			if got := shouldIssueNewCertificate(
+				testES,
+				tt.args.secret,
+				*tt.args.pod,
+				testRSAPrivateKey,
+				[]corev1.Service{testSvc},
+				testCA,
+				tt.args.rotateBefore,
+			); got != tt.want {
+				t.Errorf("shouldIssueNewCertificate() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_ensureTransportCertificatesSecretContentsForPod(t *testing.T) {
+	tests := []struct {
+		name       string
+		secret     *corev1.Secret
+		pod        *corev1.Pod
+		assertions func(t *testing.T, before corev1.Secret, after corev1.Secret)
+		wantErr    func(t *testing.T, err error)
+	}{
+		{
+			name: "no private key in the secret",
+			secret: &corev1.Secret{
+				Data: map[string][]byte{
+					PodCertFileName(testPod.Name): pemCert,
+				},
+			},
+			assertions: func(t *testing.T, before corev1.Secret, after corev1.Secret) {
+				assert.NotEmpty(t, after.Data[PodKeyFileName(testPod.Name)])
+				assert.NotEmpty(t, after.Data[PodCertFileName(testPod.Name)])
+
+				// cert should be re-generated
+				assert.NotEqual(t, after.Data[PodCertFileName(testPod.Name)], before.Data[PodCertFileName(testPod.Name)])
 			},
 		},
 		{
-			name: "should update an existing secret",
-			args: args{
-				c: k8s.WrapClient(fake.NewFakeClient(defaultPodSecretWith(func(secret *corev1.Secret) {
-					secret.ObjectMeta.UID = types.UID("42")
-				}))),
-				owner: es,
-				pod:   pod,
+			name: "no cert in the secret",
+			secret: &corev1.Secret{
+				Data: map[string][]byte{
+					PodKeyFileName(testPod.Name): certificates.EncodePEMPrivateKey(*testRSAPrivateKey),
+				},
 			},
-			want: func(t *testing.T, secret *corev1.Secret) {
-				// UID should be kept the same
-				assert.Equal(t, defaultPodSecretWith(func(secret *corev1.Secret) {
-					secret.ObjectMeta.UID = types.UID("42")
-				}), secret)
+			assertions: func(t *testing.T, before corev1.Secret, after corev1.Secret) {
+				assert.NotEmpty(t, after.Data[PodKeyFileName(testPod.Name)])
+				assert.NotEmpty(t, after.Data[PodCertFileName(testPod.Name)])
+
+				// key should be re-used
+				assert.Equal(t, before.Data[PodKeyFileName(testPod.Name)], after.Data[PodKeyFileName(testPod.Name)])
 			},
 		},
 		{
-			name: "should allow additional labels in the secret",
-			args: args{
-				c: k8s.WrapClient(fake.NewFakeClient(defaultPodSecretWith(func(secret *corev1.Secret) {
-					secret.ObjectMeta.Labels["foo"] = "bar"
-				}))),
-				owner: es,
-				pod:   pod,
+			name: "cert does not belong to the key in the secret",
+			secret: &corev1.Secret{
+				Data: map[string][]byte{
+					PodKeyFileName(testPod.Name):  certificates.EncodePEMPrivateKey(*testRSAPrivateKey),
+					PodCertFileName(testPod.Name): certificates.EncodePEMCert(testCA.Cert.Raw),
+				},
 			},
-			want: func(t *testing.T, secret *corev1.Secret) {
-				assert.Equal(t, defaultPodSecretWith(func(secret *corev1.Secret) {
-					secret.ObjectMeta.Labels["foo"] = "bar"
-				}), secret)
+			assertions: func(t *testing.T, before corev1.Secret, after corev1.Secret) {
+				assert.NotEmpty(t, after.Data[PodKeyFileName(testPod.Name)])
+				assert.NotEmpty(t, after.Data[PodCertFileName(testPod.Name)])
+
+				// key should be re-used
+				assert.Equal(t, before.Data[PodKeyFileName(testPod.Name)], after.Data[PodKeyFileName(testPod.Name)])
+				assert.NotEqual(t, after.Data[PodCertFileName(testPod.Name)], before.Data[PodCertFileName(testPod.Name)])
+			},
+		},
+		{
+			name: "invalid cert in the secret",
+			secret: &corev1.Secret{
+				Data: map[string][]byte{
+					PodKeyFileName(testPod.Name):  certificates.EncodePEMPrivateKey(*testRSAPrivateKey),
+					PodCertFileName(testPod.Name): []byte("invalid"),
+				},
+			},
+			assertions: func(t *testing.T, before corev1.Secret, after corev1.Secret) {
+				assert.NotEmpty(t, after.Data[PodKeyFileName(testPod.Name)])
+				assert.NotEmpty(t, after.Data[PodCertFileName(testPod.Name)])
+
+				// key should be re-used
+				assert.Equal(t, before.Data[PodKeyFileName(testPod.Name)], after.Data[PodKeyFileName(testPod.Name)])
+				assert.NotEqual(t, after.Data[PodCertFileName(testPod.Name)], before.Data[PodCertFileName(testPod.Name)])
+			},
+		},
+		{
+			name: "valid data should not require updating",
+			secret: &corev1.Secret{
+				Data: map[string][]byte{
+					PodKeyFileName(testPod.Name):  certificates.EncodePEMPrivateKey(*testRSAPrivateKey),
+					PodCertFileName(testPod.Name): pemCert,
+				},
+			},
+			assertions: func(t *testing.T, before corev1.Secret, after corev1.Secret) {
+				assert.Equal(t, before, after)
 			},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if tt.args.scheme == nil {
-				tt.args.scheme = scheme.Scheme
+			if tt.secret == nil {
+				tt.secret = &corev1.Secret{}
+			}
+			if tt.pod == nil {
+				tt.pod = testPod.DeepCopy()
 			}
 
-			got, err := EnsureTransportCertificateSecretExists(tt.args.c, tt.args.scheme, tt.args.owner, tt.args.pod)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("EnsureTransportCertificateSecretExists() error = %v, wantErr %v", err, tt.wantErr)
+			beforeSecret := tt.secret.DeepCopy()
+
+			err := ensureTransportCertificatesSecretContentsForPod(
+				testES,
+				tt.secret,
+				*tt.pod,
+				[]corev1.Service{testSvc},
+				testCA,
+				certificates.RotationParams{
+					Validity:     certificates.DefaultCertValidity,
+					RotateBefore: certificates.DefaultRotateBefore,
+				},
+			)
+			if tt.wantErr != nil {
+				tt.wantErr(t, err)
 				return
 			}
-			tt.want(t, got)
+			require.NoError(t, err)
+
+			tt.assertions(t, *beforeSecret, *tt.secret)
 		})
 	}
 }
