+++ operators/pkg/controller/elasticsearch/certificates/transport/reconcile.go
@@ -6,19 +6,19 @@ package transport
 
 import (
 	"bytes"
-	cryptorand "crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"fmt"
 	"reflect"
-	"time"
+	"strings"
 
 	"github.com/elastic/cloud-on-k8s/operators/pkg/apis/elasticsearch/v1alpha1"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/annotation"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/certificates"
+	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/reconciler"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/label"
+	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/name"
+	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/pod"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/utils/k8s"
 	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -27,9 +27,9 @@ import (
 
 var log = logf.Log.WithName("transport")
 
-// ReconcileTransportCertificateSecrets reconciles certificate secrets for nodes
-// of the given es cluster.
-func ReconcileTransportCertificateSecrets(
+// ReconcileTransportCertificatesSecrets reconciles the secret containing transport certificates for all nodes in the
+// cluster.
+func ReconcileTransportCertificatesSecrets(
 	c k8s.Client,
 	scheme *runtime.Scheme,
 	ca *certificates.CA,
@@ -37,7 +37,7 @@ func ReconcileTransportCertificateSecrets(
 	services []corev1.Service,
 	rotationParams certificates.RotationParams,
 ) (reconcile.Result, error) {
-	log.Info("Reconciling transport certificate secrets")
+	log.Info("Reconciling transport certificates secrets")
 
 	var pods corev1.PodList
 	if err := c.List(&client.ListOptions{
@@ -47,235 +47,127 @@ func ReconcileTransportCertificateSecrets(
 		return reconcile.Result{}, err
 	}
 
+	secret, err := ensureTransportCertificatesSecretExists(c, scheme, es)
+	if err != nil {
+		return reconcile.Result{}, err
+	}
+	// defensive copy of the current secret so we can check whether we need to update later on
+	currentTransportCertificatesSecret := secret.DeepCopy()
+
 	for _, pod := range pods.Items {
 		if pod.Status.PodIP == "" {
 			log.Info("Skipping pod because it has no IP yet", "pod", pod.Name)
 			continue
 		}
 
-		if res, err := doReconcileTransportCertificateSecret(
-			c, scheme, es, pod, services, ca, rotationParams,
+		if err := ensureTransportCertificatesSecretContentsForPod(
+			es, secret, pod, services, ca, rotationParams,
 		); err != nil {
-			return res, err
-		}
-	}
-
-	return reconcile.Result{}, nil
-}
-
-// doReconcileTransportCertificateSecret ensures that the transport certificate secret has the correct content.
-func doReconcileTransportCertificateSecret(
-	c k8s.Client,
-	scheme *runtime.Scheme,
-	es v1alpha1.Elasticsearch,
-	pod corev1.Pod,
-	svcs []corev1.Service,
-	ca *certificates.CA,
-	rotationParams certificates.RotationParams,
-) (reconcile.Result, error) {
-	secret, err := EnsureTransportCertificateSecretExists(c, scheme, es, pod)
-	if err != nil {
-		return reconcile.Result{}, err
-	}
-
-	// a placeholder secret may have nil entries, create them if needed
-	if secret.Data == nil {
-		secret.Data = make(map[string][]byte)
-	}
-	if secret.Annotations == nil {
-		secret.Annotations = make(map[string]string)
-	}
-
-	// verify that the secret contains a parsable private key, create if it does not exist
-	var privateKey *rsa.PrivateKey
-	needsNewPrivateKey := true
-	if privateKeyData, ok := secret.Data[certificates.KeyFileName]; ok {
-		storedPrivateKey, err := certificates.ParsePEMPrivateKey(privateKeyData)
-		if err != nil {
-			log.Error(err, "Unable to parse stored private key", "secret", secret.Name)
-		} else {
-			needsNewPrivateKey = false
-			privateKey = storedPrivateKey
-		}
-	}
-
-	// if we need a new private key, generate it
-	if needsNewPrivateKey {
-		generatedPrivateKey, err := rsa.GenerateKey(cryptorand.Reader, 2048)
-		if err != nil {
 			return reconcile.Result{}, err
 		}
-
-		privateKey = generatedPrivateKey
-		secret.Data[certificates.KeyFileName] = certificates.EncodePEMPrivateKey(*privateKey)
 	}
 
-	// check if the existing cert is correct
-	issueNewCertificate := shouldIssueNewCertificate(es, svcs, *secret, privateKey, ca, pod, rotationParams.RotateBefore)
-
-	if issueNewCertificate {
-		log.Info(
-			"Issuing new certificate",
-			"pod", pod.Name,
-			"es", k8s.ExtractNamespacedName(&es),
-		)
-
-		csr, err := x509.CreateCertificateRequest(cryptorand.Reader, &x509.CertificateRequest{}, privateKey)
-		if err != nil {
-			return reconcile.Result{}, err
+	// remove certificates and keys for deleted pods
+	podsByName := pod.PodsByName(pods.Items)
+	keysToPrune := make([]string, 0)
+	for secretDataKey := range secret.Data {
+		if secretDataKey == certificates.CAFileName {
+			// never remove the CA file
+			continue
 		}
 
-		// create a cert from the csr
-		parsedCSR, err := x509.ParseCertificateRequest(csr)
-		if err != nil {
-			return reconcile.Result{}, err
-		}
+		// get the pod name from the secret key name (the first segment before the ".")
+		podNameForKey := strings.SplitN(secretDataKey, ".", 2)[0]
 
-		validatedCertificateTemplate, err := CreateValidatedCertificateTemplate(pod, es, svcs, parsedCSR, rotationParams.Validity)
-		if err != nil {
-			return reconcile.Result{}, err
+		if _, ok := podsByName[podNameForKey]; !ok {
+			// pod no longer exists, so the element is safe to delete.
+			keysToPrune = append(keysToPrune, secretDataKey)
 		}
-		// sign the certificate
-		certData, err := ca.CreateCertificate(*validatedCertificateTemplate)
-		if err != nil {
-			return reconcile.Result{}, err
-		}
+	}
+	if len(keysToPrune) > 0 {
+		log.Info("Pruning keys from certificates secret", "keys", keysToPrune)
 
-		// store the issued certificate in a secret mounted into the pod
-		secret.Data[certificates.CertFileName] = certificates.EncodePEMCert(certData, ca.Cert.Raw)
+		for _, keyToRemove := range keysToPrune {
+			delete(secret.Data, keyToRemove)
+		}
 	}
 
-	// prepare trusted CA certs: CA of this node
-	trusted := certificates.EncodePEMCert(ca.Cert.Raw)
+	caBytes := certificates.EncodePEMCert(ca.Cert.Raw)
 
 	// compare with current trusted CA certs.
-	updateTrustedCACerts := !bytes.Equal(trusted, secret.Data[certificates.CAFileName])
-	if updateTrustedCACerts {
-		secret.Data[certificates.CAFileName] = trusted
+	if !bytes.Equal(caBytes, secret.Data[certificates.CAFileName]) {
+		secret.Data[certificates.CAFileName] = caBytes
 	}
 
-	if needsNewPrivateKey || issueNewCertificate || updateTrustedCACerts {
-		log.Info("Updating transport certificate secret", "secret", secret.Name)
+	if !reflect.DeepEqual(secret, currentTransportCertificatesSecret) {
 		if err := c.Update(secret); err != nil {
 			return reconcile.Result{}, err
 		}
-		annotation.MarkPodAsUpdated(c, pod)
+		for _, pod := range pods.Items {
+			annotation.MarkPodAsUpdated(c, pod)
+		}
 	}
 
 	return reconcile.Result{}, nil
 }
 
-// extractTransportCert extracts the transport certificate with the commonName from the Secret
-func extractTransportCert(secret corev1.Secret, commonName string) *x509.Certificate {
-	certData, ok := secret.Data[certificates.CertFileName]
-	if !ok {
-		log.Info("No tls certificate found in secret", "secret", secret.Name)
-		return nil
+// ensureTransportCertificatesSecretExists ensures the existence and Labels of the Secret that at a later point
+// in time will contain the transport certificates.
+func ensureTransportCertificatesSecretExists(
+	c k8s.Client,
+	scheme *runtime.Scheme,
+	es v1alpha1.Elasticsearch,
+) (*corev1.Secret, error) {
+	expected := corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: es.Namespace,
+			Name:      name.TransportCertificatesSecret(es.Name),
+
+			Labels: map[string]string{
+				// a label showing which es these certificates belongs to
+				label.ClusterNameLabelName: es.Name,
+			},
+		},
 	}
 
-	certs, err := certificates.ParsePEMCerts(certData)
-	if err != nil {
-		log.Error(err, "Invalid certificate data found, issuing new certificate", "secret", secret.Name)
-		return nil
+	// reconcile the secret resource
+	var reconciled corev1.Secret
+	if err := reconciler.ReconcileResource(reconciler.Params{
+		Client:     c,
+		Scheme:     scheme,
+		Owner:      &es,
+		Expected:   &expected,
+		Reconciled: &reconciled,
+		NeedsUpdate: func() bool {
+			// we only care about labels, not contents at this point, and we can allow additional labels
+			if reconciled.Labels == nil {
+				return true
+			}
+
+			for k, v := range expected.Labels {
+				if rv, ok := reconciled.Labels[k]; !ok || rv != v {
+					return true
+				}
+			}
+			return false
+		},
+		UpdateReconciled: func() {
+			if reconciled.Labels == nil {
+				reconciled.Labels = expected.Labels
+			} else {
+				for k, v := range expected.Labels {
+					reconciled.Labels[k] = v
+				}
+			}
+		},
+	}); err != nil {
+		return nil, err
 	}
 
-	// look for the certificate based on the CommonName
-	var names []string
-	for _, c := range certs {
-		if c.Subject.CommonName == commonName {
-			return c
-		}
-		names = append(names, c.Subject.CommonName)
+	// a placeholder secret may have nil entries, create them if needed
+	if reconciled.Data == nil {
+		reconciled.Data = make(map[string][]byte)
 	}
 
-	log.Info("Did not found a certificate with the expected common name", "secret", secret.Name, "expected", commonName, "found", names)
-
-	return nil
-}
-
-// shouldIssueNewCertificate returns true if we should issue a new certificate.
-//
-// Reasons for reissuing a certificate:
-// - no certificate yet
-// - certificate has the wrong format
-// - certificate is invalid or expired
-// - certificate SAN and IP does not match pod SAN and IP
-func shouldIssueNewCertificate(
-	cluster v1alpha1.Elasticsearch,
-	svcs []corev1.Service,
-	secret corev1.Secret,
-	privateKey *rsa.PrivateKey,
-	ca *certificates.CA,
-	pod corev1.Pod,
-	certReconcileBefore time.Duration,
-) bool {
-	certCommonName := buildCertificateCommonName(pod, cluster.Name, cluster.Namespace)
-
-	generalNames, err := buildGeneralNames(cluster, svcs, pod)
-	if err != nil {
-		log.Error(err, "Cannot create GeneralNames for the TLS certificate", "pod", pod.Name)
-		return true
-	}
-
-	cert := extractTransportCert(secret, certCommonName)
-	if cert == nil {
-		return true
-	}
-
-	publicKey, publicKeyOk := cert.PublicKey.(*rsa.PublicKey)
-	if !publicKeyOk || publicKey.N.Cmp(privateKey.PublicKey.N) != 0 || publicKey.E != privateKey.PublicKey.E {
-		log.Info(
-			"Certificate belongs do a different public key, should issue new",
-			"subject", cert.Subject,
-			"issuer", cert.Issuer,
-			"current_ca_subject", ca.Cert.Subject,
-		)
-		return true
-	}
-
-	pool := x509.NewCertPool()
-	pool.AddCert(ca.Cert)
-	verifyOpts := x509.VerifyOptions{
-		DNSName:       certCommonName,
-		Roots:         pool,
-		Intermediates: pool,
-	}
-	if _, err := cert.Verify(verifyOpts); err != nil {
-		log.Info(
-			fmt.Sprintf("Certificate was not valid, should issue new: %s", err),
-			"subject", cert.Subject,
-			"issuer", cert.Issuer,
-			"current_ca_subject", ca.Cert.Subject,
-		)
-		return true
-	}
-
-	if time.Now().After(cert.NotAfter.Add(-certReconcileBefore)) {
-		log.Info("Certificate soon to expire, should issue new", "secret", secret.Name)
-		return true
-	}
-
-	// compare actual vs. expected SANs
-	expected, err := certificates.MarshalToSubjectAlternativeNamesData(generalNames)
-	if err != nil {
-		log.Error(err, "Cannot marshal subject alternative names", "secret", secret.Name)
-		return true
-	}
-	extraExtensionFound := false
-	for _, ext := range cert.Extensions {
-		if !ext.Id.Equal(certificates.SubjectAlternativeNamesObjectIdentifier) {
-			continue
-		}
-		extraExtensionFound = true
-		if !reflect.DeepEqual(ext.Value, expected) {
-			log.Info("Certificate SANs do not match expected one, should issue new", "secret", secret.Name)
-			return true
-		}
-	}
-	if !extraExtensionFound {
-		log.Info("SAN extra extension not found, should issue new certificate", "secret", secret.Name)
-		return true
-	}
-
-	return false
+	return &reconciled, nil
 }
