+++ operators/pkg/controller/elasticsearch/volume/names.go
@@ -12,6 +12,11 @@ const (
 	KeystoreUserSecretMountPath = "/mnt/elastic-internal/keystore-user"
 	KeystoreUserVolumeName      = "elastic-internal-keystore-user"
 
+	ConfigVolumeMountPath               = "/usr/share/elasticsearch/config"
+	NodeTransportCertificatePathSegment = "node-transport-cert"
+	NodeTransportCertificateKeyFile     = "transport.tls.key"
+	NodeTransportCertificateCertFile    = "transport.tls.crt"
+
 	TransportCertificatesSecretVolumeName      = "elastic-internal-transport-certificates"
 	TransportCertificatesSecretVolumeMountPath = "/usr/share/elasticsearch/config/transport-certs"
