+++ operators/test/e2e/test/k8s_client.go
@@ -19,6 +19,7 @@ import (
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/certificates"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/certificates/http"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/name"
+	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/certificates/transport"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/label"
 	esname "github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/name"
 	kblabel "github.com/elastic/cloud-on-k8s/operators/pkg/controller/kibana/label"
@@ -232,11 +233,11 @@ func (k *K8sClient) GetCA(ownerName string, caType certificates.CAType) (*certif
 }
 
 // GetTransportCert retrieves the certificate of the CA and the transport certificate
-func (k *K8sClient) GetTransportCert(podName string) (caCert, transportCert []*x509.Certificate, err error) {
+func (k *K8sClient) GetTransportCert(esName, podName string) (caCert, transportCert []*x509.Certificate, err error) {
 	var secret corev1.Secret
 	key := types.NamespacedName{
 		Namespace: Namespace,
-		Name:      esname.TransportCertsSecret(podName),
+		Name:      esname.TransportCertificatesSecret(esName),
 	}
 	if err = k.Client.Get(key, &secret); err != nil {
 		return nil, nil, err
@@ -249,9 +250,9 @@ func (k *K8sClient) GetTransportCert(podName string) (caCert, transportCert []*x
 	if err != nil {
 		return nil, nil, err
 	}
-	transportCertBytes, exists := secret.Data[certificates.CertFileName]
+	transportCertBytes, exists := secret.Data[transport.PodCertFileName(podName)]
 	if !exists || len(transportCertBytes) == 0 {
-		return nil, nil, fmt.Errorf("no value found for secret %s", certificates.CertFileName)
+		return nil, nil, fmt.Errorf("no value found for secret %s", transport.PodCertFileName(podName))
 	}
 	transportCert, err = certificates.ParsePEMCerts(transportCertBytes)
 	if err != nil {
